The VLCTextureView class is moved to a separate file.
