+++ flutter_vlc_player/android/src/main/java/software/solid/fluttervlcplayer/FlutterVlcPlayer.java
@@ -12,13 +12,10 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.SurfaceTexture;
 import android.net.Uri;
-import android.os.Handler;
-import android.os.Looper;
 import android.util.Base64;
 import android.util.Log;
 import android.view.Surface;
 import android.view.SurfaceView;
-import android.view.TextureView;
 import android.view.View;
 
 import io.flutter.plugin.common.BinaryMessenger;
@@ -33,9 +30,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-
-import android.util.AttributeSet;
-import android.view.ViewGroup;
 
 final class FlutterVlcPlayer implements PlatformView {
 
@@ -657,233 +651,6 @@ final class FlutterVlcPlayer implements PlatformView {
         if (debug) {
             Log.d(TAG, message);
         }
-    }
-
-    // --- VLCTextureView
-
-    public class VLCTextureView extends TextureView implements TextureView.SurfaceTextureListener, View.OnLayoutChangeListener, IVLCVout.OnNewVideoLayoutListener {
-
-        private final String TAG = this.getClass().getSimpleName();
-        private boolean debug = false;
-
-        private MediaPlayer mMediaPlayer = null;
-        private TextureRegistry.SurfaceTextureEntry mTextureEntry = null;
-        protected Context mContext;
-        private SurfaceTexture mSurfaceTexture = null;
-        private boolean wasPlaying = false;
-
-        private Handler mHandler;
-        private Runnable mLayoutChangeRunnable = null;
-
-        public VLCTextureView(final Context context) {
-            super(context);
-            mContext = context;
-            initVideoView();
-        }
-
-        public VLCTextureView(final Context context, final AttributeSet attrs) {
-            super(context, attrs);
-            mContext = context;
-            initVideoView();
-        }
-
-        public VLCTextureView(Context context, AttributeSet attrs, int defStyle) {
-            super(context, attrs, defStyle);
-            mContext = context;
-            initVideoView();
-        }
-
-        public void dispose() {
-            log("************************** dispose");
-
-            setSurfaceTextureListener(null);
-            removeOnLayoutChangeListener(this);
-
-            if (mLayoutChangeRunnable != null) {
-                mHandler.removeCallbacks(mLayoutChangeRunnable);
-                mLayoutChangeRunnable = null;
-            }
-
-            if (mSurfaceTexture != null) {
-                if (!mSurfaceTexture.isReleased()) {
-                    mSurfaceTexture.release();
-                }
-                mSurfaceTexture = null;
-            }
-            mTextureEntry = null;
-            mMediaPlayer = null;
-            mContext = null;
-        }
-
-        private void initVideoView() {
-            mHandler = new Handler(Looper.getMainLooper());
-
-            setFocusable(false);
-            setSurfaceTextureListener(this);
-            addOnLayoutChangeListener(this);
-        }
-
-        public void setDebug(boolean value) {
-            debug = value;
-        }
-
-        public void setMediaPlayer(MediaPlayer mediaPlayer) {
-            log("************************** setMediaPlayer: " + mediaPlayer);
-
-            if (mediaPlayer == null) {
-                mMediaPlayer.getVLCVout().detachViews();
-            }
-
-            mMediaPlayer = mediaPlayer;
-
-            if (mMediaPlayer != null) {
-                mMediaPlayer.getVLCVout().attachViews(this);
-            }
-        }
-
-        public void setTextureEntry(TextureRegistry.SurfaceTextureEntry textureEntry) {
-            this.mTextureEntry = textureEntry;
-            this.updateSurfaceTexture();
-        }
-
-        private void updateSurfaceTexture() {
-            if (this.mTextureEntry != null) {
-                final SurfaceTexture texture = this.mTextureEntry.surfaceTexture();
-                if (!texture.isReleased() && (getSurfaceTexture() != texture)) {
-                    setSurfaceTexture(texture);
-                }
-            }
-        }
-
-        @Override
-        public void onSurfaceTextureAvailable(SurfaceTexture surface, int width, int height) {
-            log("************************** onSurfaceTextureAvailable");
-
-            if (mSurfaceTexture == null || mSurfaceTexture.isReleased()) {
-                mSurfaceTexture = surface;
-
-                if (mMediaPlayer != null) {
-                    mMediaPlayer.getVLCVout().setWindowSize(width, height);
-                    if (!mMediaPlayer.getVLCVout().areViewsAttached()) {
-                        mMediaPlayer.getVLCVout().setVideoSurface(mSurfaceTexture);
-                        if (!mMediaPlayer.getVLCVout().areViewsAttached())
-                            mMediaPlayer.getVLCVout().attachViews(this);
-                        mMediaPlayer.setVideoTrackEnabled(true);
-                        if (wasPlaying)
-                            mMediaPlayer.play();
-                    }
-                }
-                wasPlaying = false;
-
-            } else {
-                if (getSurfaceTexture() != mSurfaceTexture) {
-                    setSurfaceTexture(mSurfaceTexture);
-                }
-            }
-
-        }
-
-        @Override
-        public void onSurfaceTextureSizeChanged(SurfaceTexture surface, int width, int height) {
-            log("************************** onSurfaceTextureSizeChanged: " + width + "x" + height);
-
-            setSize(width, height);
-        }
-
-        @Override
-        public boolean onSurfaceTextureDestroyed(SurfaceTexture surface) {
-            log("************************** onSurfaceTextureDestroyed");
-
-            if (mMediaPlayer != null) {
-                wasPlaying = mMediaPlayer.isPlaying();
-            }
-
-            if (mSurfaceTexture != surface) {
-                if (mSurfaceTexture != null) {
-                    if (!mSurfaceTexture.isReleased()) {
-                        mSurfaceTexture.release();
-                    }
-                }
-                mSurfaceTexture = surface;
-            }
-
-            return false;
-        }
-
-        @Override
-        public void onSurfaceTextureUpdated(SurfaceTexture surface) {
-
-        }
-
-        @Override
-        public void onNewVideoLayout(IVLCVout vlcVout, int width, int height, int visibleWidth, int visibleHeight, int sarNum, int sarDen) {
-            log("************************** onNewLayout");
-
-            if (width * height == 0) return;
-            
-            setSize(width, height);
-        }
-
-        @Override
-        public void onLayoutChange(View view, int left, int top, int right, int bottom, int oldLeft, int oldTop, int oldRight, int oldBottom) {
-            //
-            if (left != oldLeft || top != oldTop || right != oldRight || bottom != oldBottom) {
-                log("************************** onLayoutChange: " + (oldRight - oldLeft) + "x" + (oldBottom + oldTop) + " -> " + (right - left) + "x" + (bottom - top));
-                
-                textureView.updateLayoutSize(view);
-            }
-        }
-
-        public void updateLayoutSize(View view) {
-            if (mMediaPlayer != null) {
-                mMediaPlayer.getVLCVout().setWindowSize(view.getWidth(), view.getHeight());
-                updateSurfaceTexture();
-            }
-        }
-
-        private void setSize(int width, int height) {
-            int mVideoWidth = 0;
-            int mVideoHeight = 0;
-            mVideoWidth = width;
-            mVideoHeight = height;
-            if (mVideoWidth * mVideoHeight <= 1)
-                return;
-
-            // Screen size
-            int w = this.getWidth();
-            int h = this.getHeight();
-
-            // Size
-            if (w > h && w < h) {
-                int i = w;
-                w = h;
-                h = i;
-            }
-
-            float videoAR = (float) mVideoWidth / (float) mVideoHeight;
-            float screenAR = (float) w / (float) h;
-
-            if (screenAR < videoAR)
-                h = (int) (w / videoAR);
-            else
-                w = (int) (h * videoAR);
-
-            // Layout fit
-            ViewGroup.LayoutParams lp = this.getLayoutParams();
-            lp.width = ViewGroup.LayoutParams.MATCH_PARENT;
-            lp.height = h;
-            this.setLayoutParams(lp);
-            this.invalidate();
-        }
-
-        // ---
-
-        private void log(String message) {
-            if (debug) {
-                Log.d(TAG, message);
-            }
-        }
-
     }
 
 }
